registration added, restricted some pages

+++ base/views.py
@@ -1,14 +1,21 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
 from django.contrib.auth.models import User
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Room, Topic
 from .forms import RoomForm
 
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == "POST":
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -23,11 +30,27 @@ def loginPage(request):
         else:
             messages.error(request, 'Username or password is incorrect')
 
-    return render(request, 'base/login_register.html', context={})
+    return render(request, 'base/login_register.html', context={'pageName': 'login'})
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Oops, something went wrong. Try again!')
+
+    return render(request, 'base/login_register.html', context={'pageName': 'register', 'form': form})
 
 def home(request):
     topicQuery = request.GET.get("topicQuery") if request.GET.get('topicQuery') else ''
@@ -44,7 +67,7 @@ def room(request, pk):
     room_instance = Room.objects.get(id=pk)
     return render(request, 'base/room.html', context={'room': room_instance})
 
-
+@login_required(login_url='login')
 def create_room(request):
     if request.method == "POST":
         form = RoomForm(request.POST)
@@ -54,9 +77,13 @@ def create_room(request):
 
     return render(request, 'base/room_form.html', context={'form': RoomForm()})
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room_instance = Room.objects.get(id=pk)
     form = RoomForm(instance=room_instance)
+
+    if request.user != room_instance.host:
+        return HttpResponse("<h1>You are not allowed here</h1>")
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room_instance)
@@ -66,8 +93,13 @@ def update_room(request, pk):
 
     return render(request, 'base/room_form.html', context={'form': form})
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room_instance = Room.objects.get(id=pk)
+
+    if request.user != room_instance.host:
+        return HttpResponse("<h1>You are not allowed here</h1>")
+
     if request.method == "POST":
         room_instance.delete()
         return redirect('home')
